Rename file dataclasses to energyWithCO2ePerMWh
